argus-backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.0.1
beanie==1.25.0
Brotli==1.1.0
captcha==0.5.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
coverage==7.6.1
cryptography==44.0.2
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.1.0.post1
exceptiongroup==1.2.2
fastapi==0.109.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflate64==1.0.0
iniconfig==2.1.0
lazy-model==0.2.0
Mako==1.3.9
MarkupSafe==2.1.5
minio==7.2.5
motor==2.5.1
multidict==6.1.0
multivolumefile==0.2.3
packaging==24.2
passlib==1.7.4
pillow==10.2.0
pluggy==1.5.0
propcache==0.2.0
psutil==7.0.0
py7zr==0.20.8
pyasn1==0.6.1
pybcj==1.0.2
pycparser==2.22
pycryptodome==3.22.0
pycryptodomex==3.22.0
pydantic==1.10.13
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyExifTool==0.5.6
pymongo==3.12.0
PyMySQL==1.1.0
pyppmd==1.1.0
pytest==8.3.5
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pyzstd==0.16.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.36.3
texttable==1.7.0
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.2.3
uvicorn==0.27.1
yara-python==4.5.1
yarl==1.15.2
zipp==3.20.2
